Refactor `json_player` and remove unused `json_player2.py`. Clean up redundant imports, methods, and unused logic across modules.
