+++ src/segmind/detectors/atomic_event_detectors_nodes.py
@@ -4,15 +4,10 @@ from dataclasses import dataclass, field
 from typing import Optional, List, Dict, Set, Any
 import numpy as np
 
-from giskardpy.motion_statechart.context import MotionStatechartContext
-from giskardpy.motion_statechart.data_types import ObservationStateValues
-from giskardpy.motion_statechart.graph_node import MotionStatechartNode
-from giskardpy.motion_statechart.motion_statechart import MotionStatechart
 from segmind.datastructures.events import (
     Event,
     ContactEvent,
     LossOfContactEvent,
-    MotionEvent,
     TranslationEvent,
     RotationEvent,
     StopTranslationEvent,
@@ -20,12 +15,8 @@ from segmind.datastructures.events import (
 )
 from segmind.detectors.base import (
     AbstractDetector,
-    DetectorStateChartNode,
-    SegmindContext,
 )
 from semantic_digital_twin.reasoning.predicates import contact
-from semantic_digital_twin.spatial_types.spatial_types import Pose
-from semantic_digital_twin.world_description.connections import Connection6DoF
 from semantic_digital_twin.world_description.world_entity import Body
