+++ test/test_csv_episode_segmenter.py
@@ -9,7 +9,9 @@ from unittest import TestCase
 from segmind import logger, set_logger_level, LogLevel
 import rclpy
 
+from segmind.detectors.base import SegmindContext
 from segmind.episode_segmenter import EpisodeSegmenterExecutor
+from segmind.event_logger import EventLogger
 from semantic_digital_twin.adapters.ros.visualization.viz_marker import (
     VizMarkerPublisher,
 )
@@ -44,126 +46,33 @@ class TestMultiverseEpisodeSegmenter(TestCase):
 
     @classmethod
     def setUpClass(cls):
-        logger.debug("lets go")
         multiverse_episodes_dir = (
             "/home/sorin/dev/Segmind/resources/multiverse_episodes"
         )
         selected_episode = "icub_montessori_no_hands"
         episode_dir = os.path.join(multiverse_episodes_dir, selected_episode)
         csv_file = os.path.join(episode_dir, f"data.csv")
-        print(csv_file)
-        models_dir = os.path.join(episode_dir, "models")
+        models_dir = os.path.join(episode_dir, "models/")
         cls.world = World()
+        cls.logger = EventLogger()
+        cls.context = SegmindContext(world=cls.world, logger=cls.logger)
         root = Body(name=PrefixedName(name="root", prefix="world"))
         with cls.world.modify_world():
             cls.world.add_kinematic_structure_entity(root)
-        cls.spawn_objects(models_dir)
         rclpy.init()
         cls.node = rclpy.create_node("test_node")
-        logger.debug("Node created")
         cls.viz_marker_publisher = VizMarkerPublisher(node=cls.node, _world=cls.world)
         cls.viz_marker_publisher.with_tf_publisher()
-        logger.debug("Viz marker publisher created")
         cls.file_player = CSVEpisodePlayer(
             file_path=csv_file,
             world=cls.world,
             time_between_frames=datetime.timedelta(milliseconds=4),
             position_shift=Vector3(0, 0, 0),
         )
-        logger.debug("File player created")
-        cls.episode_segmenter = NoAgentEpisodeSegmenter(
-            episode_player=cls.file_player,
-            annotate_events=True,
-            detectors_to_start=[],
-            initial_detectors=[],
+        cls.episode_segmenter = EpisodeSegmenterExecutor(player=cls.file_player, context=cls.context)
+        cls.episode_segmenter.spawn_scene(
+            models_dir=models_dir
         )
-        logger.debug("Episode segmenter created")
-
-    @classmethod
-    def spawn_objects(cls, models_dir):
-        logging.log(logging.DEBUG, f"Spawning objects from {models_dir}...")
-        # cls.copy_model_files_to_world_data_dir(models_dir)
-        directory = Path(models_dir)
-        urdf_files = [f.name for f in directory.glob("*.urdf")]
-        for file in urdf_files:
-            file_path = (
-                "/home/sorin/dev/Segmind/resources/multiverse_episodes/icub_montessori_no_hands/models/"
-                + file
-            )
-            obj_name = Path(file).stem
-
-            if obj_name == "iCub":
-                continue
-            try:
-                if obj_name == "scene":
-                    obj_world = URDFParser.from_file(file_path).parse()
-                    world_C_scene = FixedConnection(
-                        parent=cls.world.root, child=obj_world.root
-                    )
-                    with cls.world.modify_world():
-
-                        cls.world.merge_world(obj_world, world_C_scene)
-                else:
-                    obj_world = URDFParser.from_file(file_path).parse()
-                    with cls.world.modify_world():
-                        cls.world.merge_world(obj_world)
-
-            except Exception as e:
-                # import pdb
-                # pdb.set_trace()
-                logger.debug(f"Error: {e}"),
-                logger.debug(
-                    f"Could not spawn object {obj_name} from file {file}. Skipping."
-                )
-                continue
-
-    @classmethod
-    def copy_model_files_to_world_data_dir(cls, models_dir):
-        """
-        Copy the model files to the world data directory.
-        """
-        # Copy the entire folder and its contents
-        shutil.copytree(
-            models_dir, cls.world.conf.cache_dir + "/objects", dirs_exist_ok=True
-        )
-
-    @classmethod
-    def tearDownClass(cls):
-        logger.debug("Stopping the file player...")
-        logger.debug("Viz marker publisher has been stopped, exiting the world...")
-        # cls.world.exit()
-        logger.debug("World has been exited.")
-
-    def tearDown(self):
-        self.episode_segmenter.reset()
-        self.file_player.reset()
-        logger.debug("File player and episode segmenter have been reset.")
-
-    def test_containment_detector(self):
-        """
-        Test the ContainmentDetector by checking if the iCub is contained within the scene.
-        """
-        logger.debug("Testing the ContainmentDetector...")
-        self.episode_segmenter.reset()
-        self.episode_segmenter.detectors_to_start = [PlacingDetector]
-        self.episode_segmenter.initial_detectors = [
-            ContainmentDetector,
-            SupportDetector,
-        ]
-        self.episode_segmenter.start()
-
-        # self.assertTrue(any([isinstance(e, ContainmentEvent) for e in self.episode_segmenter.logger.get_events()]))
 
     def test_csv_replay(self):
-        # engine = create_engine('sqlite:///:memory:')
-        # session = Session(engine)
-        # mapper_registry.metadata.create_all(engine)
-        #
-        logger.debug("Starting the episode segmenter...")
         self.episode_segmenter.start()
-        logged_events = self.episode_segmenter.logger.get_events()
-        for l in self.episode_segmenter.logger.get_events():
-            logger.debug(f"Event: {l} at time {l.timestamp}")
-            logger.debug(10 * "-")
-        # session.add_all(self.episode_segmenter.logger.get_events())
-        # session.commit()
